Created `QuickActions` component
